Cart & CartItem are extending Checkout contracts
I want to loosen inter-module dependencies, so that a checkout doesn't
depend on a full cart, but only on a smaller subset called
CheckoutSubject and their items. This allows to use the checkout module
without the obligation to use the cart module as well. It can be useful
for shops selling a single product only.