Order contract has been extended with status, addresses and items
The package is now intentionally broken. I would want to avoid binding
this module to a concrete Address implementation, but that way it'd be
necessary either to:

1. Get rid of address methods in interface
That'd mean we have a library that has no dependencies but has no
complete functionality either.

2. Define the Order class as abstract
So that any concrete client code that uses it, would need to add their
own implementation details for billing/shipping address. Again that
would result in an incomplete library.

Either of these solutions would be appreciated by just a small niche
of use cases and the larger number of them would need to drop their
implementation.

So the solution will be rather:
- use the default vanilo address implementation
- if it's needed make it easy to opt-out and use another drop-in
  replacement for addresses

see next commits.. ;)